Refactor and clean up e2e framework utils, this patch handles test/e2e/framework/provider.go file
