+++ test/e2e/framework/BUILD
@@ -18,7 +18,6 @@ go_library(
         "nodes_util.go",
         "pods.go",
         "profile_gatherer.go",
-        "provider.go",
         "pv_util.go",
         "rc_util.go",
         "resource_usage_gatherer.go",
@@ -92,6 +91,7 @@ go_library(
         "//test/e2e/framework/metrics:go_default_library",
         "//test/e2e/framework/node:go_default_library",
         "//test/e2e/framework/pod:go_default_library",
+        "//test/e2e/framework/providers:go_default_library",
         "//test/e2e/framework/psp:go_default_library",
         "//test/e2e/framework/resource:go_default_library",
         "//test/e2e/framework/ssh:go_default_library",
@@ -139,12 +139,7 @@ filegroup(
         "//test/e2e/framework/perf:all-srcs",
         "//test/e2e/framework/pod:all-srcs",
         "//test/e2e/framework/podlogs:all-srcs",
-        "//test/e2e/framework/providers/aws:all-srcs",
-        "//test/e2e/framework/providers/azure:all-srcs",
-        "//test/e2e/framework/providers/gce:all-srcs",
-        "//test/e2e/framework/providers/kubemark:all-srcs",
-        "//test/e2e/framework/providers/openstack:all-srcs",
-        "//test/e2e/framework/providers/vsphere:all-srcs",
+        "//test/e2e/framework/providers:all-srcs",
         "//test/e2e/framework/psp:all-srcs",
         "//test/e2e/framework/replicaset:all-srcs",
         "//test/e2e/framework/resource:all-srcs",
