+++ test/e2e/framework/providers/azure/BUILD
@@ -11,6 +11,7 @@ go_library(
         "//staging/src/k8s.io/legacy-cloud-providers/azure:go_default_library",
         "//test/e2e/framework:go_default_library",
         "//test/e2e/framework/log:go_default_library",
+        "//test/e2e/framework/providers:go_default_library",
     ],
 )
