+++ test/e2e/framework/providers/gce/BUILD
@@ -26,6 +26,7 @@ go_library(
         "//test/e2e/framework/log:go_default_library",
         "//test/e2e/framework/node:go_default_library",
         "//test/e2e/framework/pod:go_default_library",
+        "//test/e2e/framework/providers:go_default_library",
         "//test/e2e/framework/service:go_default_library",
         "//test/utils:go_default_library",
         "//vendor/github.com/onsi/ginkgo:go_default_library",
