+++ test/e2e/framework/providers/kubemark/BUILD
@@ -11,6 +11,7 @@ go_library(
         "//staging/src/k8s.io/client-go/kubernetes:go_default_library",
         "//staging/src/k8s.io/client-go/tools/clientcmd:go_default_library",
         "//test/e2e/framework:go_default_library",
+        "//test/e2e/framework/providers:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
     ],
 )
