+++ test/e2e/framework/providers/vsphere/BUILD
@@ -5,7 +5,10 @@ go_library(
     srcs = ["vsphere.go"],
     importpath = "k8s.io/kubernetes/test/e2e/framework/providers/vsphere",
     visibility = ["//visibility:public"],
-    deps = ["//test/e2e/framework:go_default_library"],
+    deps = [
+        "//test/e2e/framework:go_default_library",
+        "//test/e2e/framework/providers:go_default_library",
+    ],
 )
 
 filegroup(
